Fixes in cpu/x86_ops_arith.h.

+++ src/cpu/x86_ops_arith.h
@@ -687,7 +687,7 @@ static int
 opTEST_l_a16(uint32_t fetchdat)
 {
     uint32_t temp;
-    uint16_t temp2;
+    uint32_t temp2;
 
     fetch_ea_16(fetchdat);
     if (cpu_mod != 3)
@@ -709,7 +709,7 @@ static int
 opTEST_l_a32(uint32_t fetchdat)
 {
     uint32_t temp;
-    uint16_t temp2;
+    uint32_t temp2;
 
     fetch_ea_32(fetchdat);
     if (cpu_mod != 3)
@@ -1005,7 +1005,7 @@ static int
 op83_l_a16(uint32_t fetchdat)
 {
     uint32_t src;
-    uint16_t dst;
+    uint32_t dst;
 
     fetch_ea_16(fetchdat);
     if (cpu_mod != 3)
